+++ src/typing_ex/handler/core.clj
@@ -12,10 +12,10 @@
    [ring.util.response :refer [redirect]]
    [typing-ex.boundary.drills  :as drills]
    [typing-ex.boundary.roll-calls :as roll-calls]
+   [typing-ex.boundary.restarts :as restarts]
    [typing-ex.boundary.results :as results]
    [typing-ex.boundary.stat :as stat]
-   [typing-ex.view.page :as view]
-   ))
+   [typing-ex.view.page :as view]))
 
 (comment
   (env :tp-dev)
@@ -218,3 +218,21 @@
       (if ret
         [::response/ok (str pt (get-login req) (java.util.Date.))]
         [::response/bad-request "rc! errored"]))))
+
+(defmethod ig/init-key :typing-ex.handler.core/restarts! [_ {:keys [db]}]
+  (fn [req]
+    (let [login (get-login req)
+          ret (restarts/restarts! db login)]
+      (if ret
+        [::response/ok (str "/restarts! " login)]
+        [::response/bad-request "restarts! errored"]))))
+
+(defmethod ig/init-key :typing-ex.handler.core/restarts-page [_ {:keys [db]}]
+  (fn [{[_ login] :ataraxy/result}]
+    (let [ret (restarts/restarts db login)]
+      (view/restarts-page login ret))))
+
+(defmethod ig/init-key :typing-ex.handler.core/restarts [_ {:keys [db]}]
+  (fn [{[_ login] :ataraxy/result}]
+    (let [recent (last (restarts/restarts db login))]
+      [::response/ok (str recent)])))